requirements.txt
alembic==1.13.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.4.2
pytest-asyncio==0.24.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==2.12.0

# === Runtime зависимости ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# === Pydantic v2 ===
pydantic==2.11.0
pydantic-settings==2.6.0

# === Авторизация и безопасность ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
